requirements.txt
requests
pandas
numpy
python-chess
matplotlib
seaborn
scikit-learn
streamlit
tqdm